+++ tests/pydantic_test.py
@@ -2,6 +2,7 @@
 
 from __future__ import annotations
 
+import json
 from datetime import datetime, timedelta, timezone
 
 import pytest
@@ -44,19 +45,31 @@ def test_camel_case_model() -> None:
         replace_403: bool
         foo_bar_baz: str
 
-    data = TestModel.parse_obj(
-        {"minimumLifetime": 10, "replace403": False, "fooBarBaz": "something"}
-    )
+    camel = {
+        "minimumLifetime": 10,
+        "replace403": False,
+        "fooBarBaz": "something",
+    }
+    snake = {
+        "minimum_lifetime": 10,
+        "replace_403": False,
+        "foo_bar_baz": "something",
+    }
+    data = TestModel.parse_obj(camel)
     assert data.minimum_lifetime == 10
     assert not data.replace_403
     assert data.foo_bar_baz == "something"
+    assert data.dict() == camel
+    assert data.dict(by_alias=False) == snake
+    assert data.json() == json.dumps(camel)
+    assert data.json(by_alias=False) == json.dumps(snake)
 
-    data = TestModel.parse_obj(
-        {"minimum_lifetime": 20, "replace_403": True, "foo_bar_baz": "else"}
-    )
-    assert data.minimum_lifetime == 20
-    assert data.replace_403
-    assert data.foo_bar_baz == "else"
+    snake_data = TestModel.parse_obj(snake)
+    assert data.minimum_lifetime == 10
+    assert not data.replace_403
+    assert data.foo_bar_baz == "something"
+    assert snake_data.dict() == data.dict()
+    assert snake_data.json() == data.json()
 
 
 def test_validate_exactly_one_of() -> None: